+++ src/cmd/go/internal/doc/pkgsite_bootstrap.go
@@ -4,10 +4,8 @@
 
 //go:build cmd_go_bootstrap
 
-// Don't build cmd/doc into go_bootstrap because it depends on net.
+// Don't build the pkgsite code into go_bootstrap because it depends on net.
 
 package doc
 
-import "cmd/go/internal/base"
-
-var CmdDoc = &base.Command{}
+func doPkgsite(string) error { return nil }
